refactor: remove unused chat_room_id from users
I added this when I thought I needed to keep track of which chat room a user was in through a relationship in the database. With the addition of the websocket connection we don't need to keep track of it in this way, rather we can allow that connection to be the relationship.